DB永続化

+++ src/models.rs
@@ -1,8 +1,9 @@
 use serde::{Serialize, Deserialize};
+use sqlx::FromRow;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct Todo {
-  pub id: u32,
+  pub id: i64,
   pub title: String,
   pub completed: bool,
 }
